% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{annoExonTable}
\alias{annoExonTable}
\title{Build exon table from annotated isoforms}
\usage{
annoExonTable(gene_bed, gtf, gene, split = "|")
}
\arguments{
\item{gene_bed}{A dataframe of a bed annotation.}

\item{gtf}{A dataframe of gtf annotation.}

\item{gene}{The target gene to build annotated exon table}

\item{split}{A charater used to split the exon ids in the isoform}
}
\value{
A data frame for the exon count in each annotated isoform,
each row is an isoform, each col is an exon.
}
\description{
Count the exon incorporation for each annotated isoform
}
\details{
Input annotated isoforms and return a table of exon count for each isoform
}
