% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom_mle.R
\name{bbmoment}
\alias{bbmoment}
\title{bbmoment}
\usage{
bbmoment(success, total)
}
\arguments{
\item{success}{a vector for the number of success}

\item{total}{a vector for the number of total trials}
}
\value{
a vector include the moment estimation for alpha and beta
}
\description{
moment estimation for alpha and beta parameters in beta-binomial
}
