% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{beta_dis}
\alias{beta_dis}
\title{distance between two psi distribution modeled by beta.}
\usage{
beta_dis(
  exon_count1,
  gene_count1,
  exon_count2,
  gene_count2,
  iters = 100,
  psi_num = 500
)
}
\arguments{
\item{exon_count1}{a vector of exon count for cell population 1}

\item{gene_count1}{a vector of gene count for cell population 1}

\item{exon_count2}{a vector of exon count for cell population 2}

\item{gene_count2}{a vector of gene count for cell population 2}

\item{iters}{the number of iterations for bootstrap}

\item{psi_num}{the sample size from two beta distributions to calculate the earth-moving distance}
}
\value{
a vector of the distance and its confidence interval
}
\description{
for significant difference of alternative splicing, this function calculates
the earth-moving distance between the psi distribution of two cell populations and its confidence interval
through bootstrap.
}
