% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{beta_mean}
\alias{beta_mean}
\title{calculate mean value for a beta distribution}
\usage{
beta_mean(alpha, beta)
}
\arguments{
\item{alpha}{alpha for the beta distribution}

\item{beta}{beta for the beta distribution}
}
\value{
a numeric value as the mean
}
\description{
calculate mean value for a beta distribution
}
