% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{beta_var}
\alias{beta_var}
\title{calculate variance for a beta distribution}
\usage{
beta_var(alpha, beta)
}
\arguments{
\item{alpha}{alpha for the beta distribution}

\item{beta}{beta for the beta distribution}
}
\value{
a numeric value as the variance
}
\description{
calculate variance for a beta distribution
}
