% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{cellGeneExonCount}
\alias{cellGeneExonCount}
\title{count exons in each cell}
\usage{
cellGeneExonCount(exon_table, cell_col = "cell_id")
}
\arguments{
\item{exon_table}{a dataframe of count for mutil-exons in each isoform}

\item{cell_col}{the name of the cell column in the exon table}
}
\value{
a list of exon count table and gene count table for each cell
}
\description{
summarise the exon count and gene count for each single cell
}
\details{
Due to the truncation, the exons are counted in a censoring way, only
exons between two exons or between an exon and polyA will be treated as spliced out,
otherwise it is censor and won't be counted.
}
