% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{cellGeneSiteCount}
\alias{cellGeneSiteCount}
\title{cellGeneSiteCount}
\usage{
cellGeneSiteCount(splice_site_table, cell_col = "cell", count_col = "count")
}
\arguments{
\item{splice_site_table}{a dataframe of count for splicing sites in each isoform}

\item{cell_col}{The name of the column recording the cell barcode}

\item{count_col}{The name of the column recording the isoform count in each cell}
}
\value{
a list of site count table and gene count table for each cell

A list including two dataframes, the fisrt one records the gene count for each site in each cell,
while the second one record the spliced-in site count.
}
\description{
summarise the site count and gene count for each single cell
}
\details{
Due to the truncation, the sites are counted in a censoring way, only
sites between two sites or between a site and the polyA will be treated as spliced out,
otherwise it is censor and won't be counted.
}
