% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{createSpliceObject}
\alias{createSpliceObject}
\title{Create a Splice object}
\usage{
createSpliceObject(
  path,
  cell = "all",
  gene = "all",
  cell_col = "cell",
  gene_col = "gene",
  iso_col = "exons"
)
}
\arguments{
\item{path}{The path to store the output files from saveExonList().}

\item{cell}{The vector of cells to be preserved in the Splice object}

\item{gene}{The vector of genes to be preserved in the Splice object}

\item{cell_col}{The colname to for the cell id in the count file}

\item{gene_col}{The colname to for the gene id in the count file}

\item{iso_col}{The colname to for the isoform id in the count file}
}
\description{
Create a Splice object from the output file from saveExonList().
}
\details{
Create a Splice object from the output file from saveExonList().
}
