% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{createSpliceSeurat}
\alias{createSpliceSeurat}
\title{Create a Seurat object with Splice object}
\usage{
createSpliceSeurat(
  path,
  project = "SeuratProject",
  assay = "RNA",
  min.cells = 3,
  min.features = 20,
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)
}
\arguments{
\item{path}{The path to store the output files from saveExonList()}

\item{project}{The name of the project for this object}

\item{assay}{The assay name for the count matrix}

\item{min.cells}{The minimum number of cells to express a gene}

\item{min.features}{The minimum number of genes cell to express}

\item{names.field}{For the initial identity class for each cell, choose this field from the cell's name.}

\item{names.delim}{For the initial identity class for each cell, choose this delimiter from the cell's column name}

\item{meta.data}{Additional cell-level metadata to add to the Seurat object.}

\item{...}{other parameters for Seurat::CreateSeuratObject}
}
\description{
Create a Seurat and Splice object from the output file from saveExonList().
}
\details{
The Seraut object is created to store the gene expression information,
and the Splice object for isoform information. The Splice object is embedded in the
misc slot in Seurat object
}
