% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{creatSplice_from_df}
\alias{creatSplice_from_df}
\title{creatSplice_from_df}
\usage{
creatSplice_from_df(
  df,
  cell = "all",
  gene = "all",
  cell_col = "cell",
  gene_col = "gene",
  iso_col = "isoform",
  exprs_col = "count"
)
}
\arguments{
\item{df}{The input dataframe.}

\item{cell}{The vector of cells to be preserved in the Splice object}

\item{gene}{The vector of genes to be preserved in the Splice object}

\item{cell_col}{The colname to for the cell id in the count file}

\item{gene_col}{The colname to for the gene id in the count file}

\item{iso_col}{The colname to for the isoform id in the count file}
}
\value{
A Splice object
}
\description{
Create a Splice object from the output file from a dataframe.
}
\details{
Create a Splice object from the output file from dataframe, the dataframe should
contain at least four columns, including cell, gene, isoform and umi count.
}
