% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{dataAnnoCombine}
\alias{dataAnnoCombine}
\title{combine the table for exon merge}
\usage{
dataAnnoCombine(a, b)
}
\arguments{
\item{a}{exon merge flag table from data}

\item{b}{exon merge flag table from annotation}
}
\value{
A square matrix of logical values, indicating if each two exons should be merged
}
\description{
combine merge table from data(a) and annotation(b)
}
\details{
Exons are merged according to their co-existence or mutual exclusiveness in isoforms.
Isoforms in data can be different in annotations, this function will choose the intersection for
the merge flag
}
