% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exon_corres}
\alias{exon_corres}
\title{assign canonical exons to self annotated sub-exons}
\usage{
exon_corres(
  start,
  end,
  bed,
  exon_id = NULL,
  strand_col = "strand",
  start_col = "start",
  end_col = "end"
)
}
\arguments{
\item{start, end}{The start and end site of the canonical exon}

\item{bed}{self annotated bed for a gene}

\item{exon_id}{A vector of string to name the self annotated exon, should have the same length
as the self-annotated exons}

\item{strand_col}{The colname to indicate the strand in the self-annotated gene bed}

\item{start_col, end_col}{The colnames of start and end sites in the self-annotated gene bed}
}
\value{
A vector of sub-exons contained by the canonical exon.
}
\description{
assign canonical exons to self annotated sub-exons
}
