% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exon_len_filter}
\alias{exon_len_filter}
\title{filter out too short exons}
\usage{
exon_len_filter(
  gene,
  gene_bed,
  thresh = 10,
  exons = NULL,
  exon_id = NULL,
  gene_id = "gene",
  len = "width",
  strand = "strand",
  sep = "|"
)
}
\arguments{
\item{gene}{name of the target gene}

\item{gene_bed}{gene bed annotation}

\item{thresh}{the minimum length for the exon to be}

\item{exons}{selected exons, exon filtering will be within this selection}

\item{exon_id}{the vector of names for each exon}

\item{gene_id}{the name of gene column in the bed file}

\item{len}{the name of exon length column in the bed file}

\item{strand}{the name of strand column in the bed file}

\item{sep}{The seperation character between exons in the isoform string}
}
\value{
a vector of preserved exons
}
\description{
filter out too short exons which can be confused with insertions or deletions
}
\details{
Some sub-exons are too short and their existence can be confused by insertions
or deletions, which can misled the downstream alternative splicing analysis.
}
