% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exon_search}
\alias{exon_search}
\title{Search if an exon is within an exon group.}
\usage{
exon_search(exon, exons_group, sep = "|")
}
\arguments{
\item{exon}{The specified exon to be observed.}

\item{exons_group}{A vector of strings, each one is a meta-exon}

\item{sep}{The separator between each exon in the meta-exon}
}
\value{
A numerical value representing the id of the meta-exon which incorporated the specified
exon.
}
\description{
Search if an exon is within an exon group.
}
\details{
Some exons can be merged into a meta-exon due to co-existence or mutually exclusiveness,
when such exons are required to be observed, we need to search them in the meta-exon.
}
