% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonExonMerge}
\alias{exonExonMerge}
\title{merge two exons}
\usage{
exonExonMerge(a, b)
}
\arguments{
\item{a}{count vector for exon a}

\item{b}{count vector for exon b}
}
\value{
a count vector for merged exon
}
\description{
merge the existence of two exons in each isoform
}
\details{
Input the count vectors for two exons in each isoform, cencoring in one exon can
be recovered by the information from the other
}
