% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonMatrixMerge}
\alias{exonMatrixMerge}
\title{combine the merge flag from data and annotation}
\usage{
exonMatrixMerge(dataExonMergeMatrix, annoExonMergeMatrix = NULL)
}
\arguments{
\item{dataExonMergeMatrix}{exon merge flag matrix generated from data}

\item{annoExonMergeMatrix}{exon merge flag matrix generated from annotation}
}
\value{
A list of vectors with variable length,each vector represents a group of exons to be merged
}
\description{
combine the merge flag from data and annotation
}
\details{
Input the exon merge flag matrix generated from data and annotation and output
exon groups to be merged
}
