% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonTable}
\alias{exonTable}
\title{Build exon table from isoforms}
\usage{
exonTable(exons, count, polyA, split = "|")
}
\arguments{
\item{exons}{A vector of string representing isoform, should be a sequence of exon id.}

\item{count}{A numerical vector representing the count for each isoform}

\item{polyA}{A numerical vector representing the polyA existence for each isoform}

\item{split}{A charater used to split the exon ids in the isoform}
}
\value{
A data frame for the exon count in each isoform,
each row is an isoform, each col is an exon.
}
\description{
Count the exon incorporation for each isoform
}
\details{
Input isoforms and return a table of exon count for each isoform in cencoring way
}
