% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonTableMerge}
\alias{exonTableMerge}
\title{merge exons}
\usage{
exonTableMerge(exon_table, gene_bed = NULL, gtf = NULL, gene = NULL, sep = "|")
}
\arguments{
\item{exon_table}{a dataframe of count for mutil-exons in each isoform}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{gene}{the gene to refer annotation}

\item{sep}{The seperation character between exons in the isoform string}
}
\value{
a count matrix for merged exon
}
\description{
merge exons according to their relationship, co-existing or mutual exclusive
}
\details{
Input a dataframe of count for mutil-exons in each isoform and the bed
and gtf annotation to identify the relationship between exon pairs. Exon sets which are
agreed to be merged by both data and annotation are merged
}
