% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonTableMergeFlag}
\alias{exonTableMergeFlag}
\title{judge if exons can be merged}
\usage{
exonTableMergeFlag(exon_table)
}
\arguments{
\item{exon_table}{The exon count table output from exonTable()}
}
\value{
A square matrix of logical values, indicating if each two exons should be merged
}
\description{
judge if each exon in the gene can be merged with other
}
\details{
Iteratively check if each pair of exons in the gene co-existing or
mutually exclusive
}
