% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{exonTableMergeList}
\alias{exonTableMergeList}
\title{merge exons in an exon table according to to be merged list}
\usage{
exonTableMergeList(exon_table, merge_list, sep = "|")
}
\arguments{
\item{exon_table}{a dataframe of count for mutil-exons in each isoform}

\item{merge_list}{A list of vectors with variable length,each vector represents a group of exons to be merged}

\item{sep}{The seperation character between exons in the isoform string}
}
\value{
a count vector for merged exon
}
\description{
merge exons in an exon table according to to be merged list
}
\details{
Input a dataframe of count for mutil-exons in each isoform and a list
indicating which exons should be merged together. Exons in the same group will be
merged into a meta-exon
}
