% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{gene_exons_phi}
\alias{gene_exons_phi}
\title{compute phi for exons in a gene}
\usage{
gene_exons_phi(
  spliceOb,
  gene,
  gene_bed = NULL,
  gtf = NULL,
  cells = "all",
  exons = "all",
  exon_len_thresh = 10,
  cell_count_thresh = 30,
  exprs_thresh = 10,
  iters = 100,
  sep = "|"
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{cells}{cells to use in building exon count table}

\item{exons}{exon to use in building exon count table}

\item{exon_len_thresh}{the minimum length of an exon to be preserved in exon count table}

\item{cell_count_thresh}{the minimum number of cells expressing this gene}

\item{exprs_thresh}{the minimum gene expression to calculate the exon psi (for unbiased estimation)}

\item{iters}{iteration for bootstrap to generate the confidence interval of phi}

\item{sep}{The seperation character between exons in the isoform string}
}
\value{
a list of exon count table and gene count table for each cell
}
\description{
compute phi for exons in a gene from a Splice object
}
