% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{gene_exons_psi,Seurat,character,character,character-method}
\alias{gene_exons_psi,Seurat,character,character,character-method}
\title{generic gene_exons_psi function for Seurat object}
\usage{
\S4method{gene_exons_psi}{Seurat,character,character,character}(
  object,
  gene,
  cells = "all",
  exons = "all",
  slot = "splice",
  ...
)
}
\arguments{
\item{object}{A Seurat object embedded with a Splice object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{exons}{exon to use in building exon count table}

\item{slot}{the slot name for the Splice object embedded in the Seurat object}

\item{...}{other possible parameters for gene_exons_psi.base}
}
\description{
generic gene_exons_psi function for Seurat object
}
