% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{gene_exons_psi}
\alias{gene_exons_psi}
\title{generic genes_exons_psi function definition}
\usage{
gene_exons_psi(object, gene, cells, exons, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{exons}{exon to use in building exon count table}

\item{...}{other possible parameters for gene_exons_psi.base}
}
\description{
generic genes_exons_psi function definition
}
