% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{gene_exons_table,Seurat,character-method}
\alias{gene_exons_table,Seurat,character-method}
\title{generic gene_exons_table function for Seurat object}
\usage{
\S4method{gene_exons_table}{Seurat,character}(
  object,
  gene,
  slot = "splice",
  gene_bed = NULL,
  gtf = NULL,
  ...
)
}
\arguments{
\item{object}{a Seurat object with a Splice object in misc}

\item{gene}{name of the target gene}

\item{slot}{the slot name for the Splice object embedded in the Seurat object}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{...}{parameters for gene_exons_table.base}
}
\description{
generic gene_exons_table function for Seurat object
}
