% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{gene_exons_table,Splice,character-method}
\alias{gene_exons_table,Splice,character-method}
\title{generic gene_exons_table function for Splice object}
\usage{
\S4method{gene_exons_table}{Splice,character}(object, gene, gene_bed = NULL, gtf = NULL, ...)
}
\arguments{
\item{object}{a Splice object}

\item{gene}{name of the target gene}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{...}{parameters for gene_exons_table.base}
}
\description{
generic gene_exons_table function for Splice object
}
