% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{gene_exons_table.base}
\alias{gene_exons_table.base}
\title{count exons in each cell from a Splice object}
\usage{
gene_exons_table.base(
  spliceOb,
  gene,
  gene_bed = NULL,
  gtf = NULL,
  cells = "all",
  exons = "all",
  exon_len_thresh = 10,
  sep = "|"
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{cells}{cells to use in building exon count table}

\item{exons}{exon to use in building exon count table}

\item{exon_len_thresh}{the minimum length of an exon to be preserved in exon count table}

\item{sep}{The seperation character between exons in the isoform string}
}
\value{
a list of exon count table and gene count table for each cell
}
\description{
summarise the exon count and gene count for each single cell from a Splice object
}
\details{
Due to the truncation, the exons are counted in a censoring way, only
exons between two exons or between an exon and polyA will be treated as spliced out,
otherwise it is censor and won't be counted.
}
