% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{gene_exons_table}
\alias{gene_exons_table}
\title{generic gene_exons_table function definition}
\usage{
gene_exons_table(object, gene, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene}{name of the target gene}

\item{...}{other possible parameters for genes_groups_GLRT.base}
}
\description{
generic gene_exons_table function definition
}
