% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{gene_GLRT,Seurat,character,character,character-method}
\alias{gene_GLRT,Seurat,character,character,character-method}
\title{generic gene_GLRT function for Seurat object}
\usage{
\S4method{gene_GLRT}{Seurat,character,character,character}(object, gene, group1, group2, slot = "splice", group_col = NULL, ...)
}
\arguments{
\item{object}{the Seurat object}

\item{gene}{the name of specified gene}

\item{group1}{group name or cell name}

\item{group2}{group name or cell name}

\item{slot}{the slot name to store Splice object in Seurat object}

\item{group_col}{if group_col is NULL, group1 and group2 should be cell names in
each group. If group_col is specified as a colname in Seurat cell meta data,
group1 and group2 should be group names.}

\item{...}{parameters specified for gene_exons_table.base()}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for specified exons in a gene
for specified two cell groups from a Seurat object.
}
