% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{gene_GLRT,Splice,character,vector,vector-method}
\alias{gene_GLRT,Splice,character,vector,vector-method}
\title{generic gene_GLRT function for Splice object}
\usage{
\S4method{gene_GLRT}{Splice,character,vector,vector}(object, gene, group1, group2, ...)
}
\arguments{
\item{object}{the Splice object}

\item{gene}{the name of specified gene}

\item{group1}{cells for group 1}

\item{group2}{cells for group 2}

\item{...}{parameters specified for gene_exons_table.base()}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for specified exons in a gene
for specified two cell groups from a Splice object.
}
