% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{gene_GLRT.base}
\alias{gene_GLRT.base}
\title{Generalized likelihood ratio test for a gene.}
\usage{
gene_GLRT.base(
  spliceOb,
  gene,
  group1,
  group2,
  exons = "all",
  iters = 100,
  psi_num = 500,
  group_cell_thresh = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{group1}{cells for group 1}

\item{group2}{cells for group 2}

\item{exons}{specified exons to do GLRT}

\item{iters}{the number of iterations for bootstrap}

\item{psi_num}{the sample size from two beta distributions to calculate
the earth-moving distance}

\item{group_cell_thresh}{the minimum cell number for each group for GLRT}

\item{verbose}{whether to print out the iteration information}

\item{...}{parameters specified for geneSiteTable()}

\item{gene_bed}{a dataframe of gene bed annotation}

\item{gtf}{a dataframe of gtf annotation}

\item{exon_len_thresh}{the minimum length of exons for GLRT}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for specified exons in a gene
for specified two cell groups from a Splice object.
}
