% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{gene_GLRT}
\alias{gene_GLRT}
\title{generic gene_GLRT function definition}
\usage{
gene_GLRT(object, gene, group1, group2, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene}{the name of specified gene}

\item{group1}{cells for group 1}

\item{group2}{cells for group 2}

\item{...}{other possible parameters for gene_GLRT.base}
}
\description{
generic gene_GLRT function definition
}
