% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{genes_exons_phi,Seurat,character-method}
\alias{genes_exons_phi,Seurat,character-method}
\title{generic genes_exons_phi function for Seurat object}
\usage{
\S4method{genes_exons_phi}{Seurat,character}(object, genes, slot = "splice", ...)
}
\arguments{
\item{object}{The Seurat object with a Splice object in misc}

\item{genes}{vector of names of the target gene}

\item{slot}{the slot name for the Splice object embedded in the Seurat object}

\item{...}{parameters for genes_exons_phi.base}
}
\description{
generic genes_exons_phi function for Seurat object
}
