% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{genes_exons_phi,Splice,character-method}
\alias{genes_exons_phi,Splice,character-method}
\title{generic genes_exons_phi function for Splice object}
\usage{
\S4method{genes_exons_phi}{Splice,character}(object, genes, ...)
}
\arguments{
\item{object}{the Splice object}

\item{genes}{vector of names of the target gene}

\item{...}{parameters for genes_exons_phi.base}
}
\description{
generic genes_exons_phi function for Splice object
}
