% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{genes_exons_phi.base}
\alias{genes_exons_phi.base}
\title{compute phi for exons in multi-genes}
\usage{
genes_exons_phi.base(
  spliceOb,
  genes,
  cells = "all",
  exons = "all",
  gene_bed = NULL,
  gtf = NULL,
  exon_len_thresh = 10,
  cell_count_thresh = 30,
  exprs_thresh = 10,
  iters = 100,
  psi_lwr = 0.1,
  psi_upr = 0.9,
  phi_conf_thresh = 0.2,
  sep = "|",
  cores = 4
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{genes}{vector of names of the target gene}

\item{cells}{cells to use in building exon count table}

\item{exons}{exon to use in building exon count table}

\item{gene_bed}{gene bed annotation}

\item{gtf}{gtf annotation}

\item{exon_len_thresh}{the minimum length of an exon to be preserved in exon count table}

\item{cell_count_thresh}{the minimum number of cells expressing this gene}

\item{exprs_thresh}{the minimum gene expression to calculate the exon psi (for unbiased estimation)}

\item{iters}{The number of iterations for bootstrapping to generate confidence interval}

\item{psi_lwr, psi_upr}{The lower and upper bound to preserve as alternative spliced exons}

\item{phi_conf_thresh}{The maximum length of confidence interval to filter out inconfident
phi estimation}

\item{sep}{The seperation character between exons in the isoform string}

\item{cores}{The number of cores for parallization}
}
\value{
a list of exon count table and gene count table for each cell
}
\description{
compute phi for exons in a gene from a Splice object
}
