% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{genes_exons_phi}
\alias{genes_exons_phi}
\title{generic genes_exons_phi function definition}
\usage{
genes_exons_phi(object, genes, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{genes}{vector of names of the target gene}

\item{...}{other possible parameters}
}
\description{
generic genes_exons_phi function definition
}
