% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{genes_groups_GLRT,Seurat,character,character,character-method}
\alias{genes_groups_GLRT,Seurat,character,character,character-method}
\title{generic genes_groups_GLRT function for Seurat object}
\usage{
\S4method{genes_groups_GLRT}{Seurat,character,character,character}(
  object,
  genes,
  group1s,
  group2s,
  group_col = NULL,
  slot = "splice",
  ...
)
}
\arguments{
\item{object}{the Seurat object}

\item{genes}{a vector of gene names to do GLRT}

\item{group1s}{group names or cell names}

\item{group2s}{group names or cell names}

\item{group_col}{if group_col is NULL, group1s and group2s should be cell names in
each group. If group_col is specified as a colname in Seurat cell meta data,
group1s and group2s should be group names.}

\item{slot}{the slot name to store Splice object in Seurat object}

\item{...}{parameters for genes_groups_GLRT.base}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for all exons in multiple genes
for multiple cell group pairs from a Splice object.
}
