% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{genes_groups_GLRT,Splice,character,list,list-method}
\alias{genes_groups_GLRT,Splice,character,list,list-method}
\title{generic genes_groups_GLRT function for Splice object}
\usage{
\S4method{genes_groups_GLRT}{Splice,character,list,list}(object, genes, group1s, group2s, ...)
}
\arguments{
\item{object}{the Splice object}

\item{genes}{a vector of gene names to do GLRT}

\item{group1s}{a named list of cell names, each element should be cells belonging to a group}

\item{group2s}{a named list of cell names, each element should be cells belonging to a group}

\item{...}{other parameters for gene_GLRT.base}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for all exons in multiple genes
for multiple cell group pairs from a Splice object.
}
