% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{genes_groups_GLRT.base}
\alias{genes_groups_GLRT.base}
\title{Generalized likelihood ratio test for multiples gene.}
\usage{
genes_groups_GLRT.base(
  spliceOb,
  genes,
  group1s,
  group2s,
  iters = 100,
  psi_num = 500,
  group_cell_thresh = 30,
  q_thresh = 0.05,
  cores = 4,
  filter = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{genes}{a vector of gene names to do GLRT}

\item{group1s}{a named list of cell names, each element should be cells belonging to a group}

\item{group2s}{a named list of cell names, each element should be cells belonging to a group}

\item{iters}{the number of iterations for bootstrap}

\item{psi_num}{the sample size from two beta distributions to calculate
the earth-moving distance}

\item{group_cell_thresh}{the minimum cell number for each group for GLRT}

\item{q_thresh}{the threshold for q value by fdr control}

\item{cores}{The number of cores used for parallization}

\item{verbose}{if print the process information}

\item{...}{other parameters for gene_GLRT.base}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
an interface function to run GLRT for all exons in multiple genes
for multiple cell group pairs from a Splice object.
}
