% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{genes_groups_GLRT}
\alias{genes_groups_GLRT}
\title{generic genes_groups_GLRT function definition}
\usage{
genes_groups_GLRT(object, genes, group1s, group2s, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{genes}{a vector of gene names to do GLRT}

\item{group1s}{a named list of cell names, each element should be cells belonging to a group}

\item{group2s}{a named list of cell names, each element should be cells belonging to a group}

\item{...}{other possible parameters for genes_groups_GLRT.base}
}
\description{
generic genes_groups_GLRT function definition
}
