% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{genes_multigroups_GLRT}
\alias{genes_multigroups_GLRT}
\title{genes_multigroups_GLRT}
\usage{
genes_multigroups_GLRT(
  spliceOb,
  genes,
  meta,
  cell_col = "cell",
  group_col = "group",
  filter = TRUE,
  q_thresh = 0.05,
  ...
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{genes}{target genes to be tested}

\item{meta}{a dataframe recording the meta data for each, should have at lease two columns,
one indicate the cell barcode and the other one indicate the cell group information.}

\item{cell_col, group_col}{The column names in the meta to record the cell/group info.}

\item{filter}{A bool flag to indicate if insignificant results should be filtered out.}

\item{q_thresh}{The threshold for the FDR controled q value.}

\item{...}{Other parameters used in gene_GLRT.base.}
}
\value{
a dataframe of the test result. Each line is an exon with p-values
and parameters for the psi distributions in two cell populations if they
have significant difference
}
\description{
Do the generalized likelihood ratio test for multiples gene across multiple cell groups.
}
