% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{geneSitesPhi_from_beta}
\alias{geneSitesPhi_from_beta}
\title{geneSitesPhi_from_beta}
\usage{
geneSitesPhi_from_beta(
  spliceOb,
  gene,
  cells = "all",
  mean_psi_lwr = 0.01,
  mean_psi_upr = 0.99,
  cell_count_thresh = 20,
  exprs_thresh = 5,
  CI_level = 0.05
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{cell_count_thresh}{the minimum number of cells expressing this gene}

\item{exprs_thresh}{the minimum gene expression to calculate the exon psi (for unbiased estimation)}

\item{CI_level}{The alpha for the confidence interval of phi estimation}

\item{exon_len_thresh}{the minimum length of an exon to be preserved in exon count table}
}
\value{
a dataframe recording the phi estimation
}
\description{
compute phi for exons in a gene from a Splice object
}
