% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{geneSitesPhi_from_psi}
\alias{geneSitesPhi_from_psi}
\title{geneSitesPhi_from_psi}
\usage{
geneSitesPhi_from_psi(
  spliceOb,
  gene,
  cells = "all",
  cell_count_thresh = 30,
  exprs_thresh = 10,
  iters = 100,
  ...
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{cell_count_thresh}{the minimum number of cells expressing this gene}

\item{exprs_thresh}{the minimum gene expression to calculate the exon psi (for unbiased estimation)}

\item{iters}{iteration for bootstrap to generate the confidence interval of phi}

\item{...}{Other parameters for geneSitesPsi.base}
}
\value{
a dataframe recording the phi estimation
}
\description{
compute phi for exons in a gene from a Splice object
}
