% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{geneSitesPsi,Splice,character-method}
\alias{geneSitesPsi,Splice,character-method}
\title{generic geneSitesPsi function for Splice object}
\usage{
\S4method{geneSitesPsi}{Splice,character}(object, gene, ...)
}
\arguments{
\item{object}{A Splice object}

\item{gene}{name of the target gene}

\item{...}{other possible parameters for geneSitesPsi.base}
}
\description{
generic geneSitesPsi function for Splice object
}
