% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{geneSitesPsi.base}
\alias{geneSitesPsi.base}
\title{geneSitesPsi.base}
\usage{
geneSitesPsi.base(spliceOb, gene, cells = "all", exprs_thresh = 10, ...)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{exprs_thresh}{The minimum threshold for the gene expression to estimate the psi}
}
\value{
a dataframe contains the psi estimation for each exon in each cell
}
\description{
compute psi for sites in a gene from a Splice object
}
