% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{geneSitesPsi}
\alias{geneSitesPsi}
\title{generic genes_exons_psi function definition}
\usage{
geneSitesPsi(object, gene, ...)
}
\arguments{
\item{object}{the Splice object}

\item{gene}{name of the target gene}

\item{...}{other possible parameters for geneSitesPsi.base}
}
\description{
generic genes_exons_psi function definition
}
