% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{geneSiteTable,Splice,list,character-method}
\alias{geneSiteTable,Splice,list,character-method}
\title{generic gene_exons_table function for Splice object}
\usage{
\S4method{geneSiteTable}{Splice,list,character}(object, gene_bed, genes, eps, ...)
}
\arguments{
\item{object}{a Splice object}

\item{gene_bed}{The gene bed annotation, should include the gene name and the gene strand.}

\item{genes}{The target genes to generate meta sites, default for all genes}

\item{eps}{The threshold for two sites to be merged}

\item{...}{parameters for geneSiteTable.base}
}
\description{
generic gene_exons_table function for Splice object
}
