% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{geneSiteTable.base}
\alias{geneSiteTable.base}
\title{geneSiteTable.base}
\usage{
geneSiteTable.base(
  spliceOb,
  gene_bed,
  genes = "all",
  eps = 0.05,
  bed_gene_col = "gene",
  bed_strand_col = "strand",
  verbose = TRUE
)
}
\arguments{
\item{spliceOb}{A Splice object}

\item{gene_bed}{The gene bed annotation, should include the gene name and the gene strand.}

\item{genes}{The target genes to generate meta sites, default for all genes}

\item{eps}{The threshold for two sites to be merged}

\item{bed_gene_col}{The name of the column to record the gene name in the gene bed}

\item{bed_strand_col}{The name of the column to record the gene strand in the gene bed}

\item{verbose}{The flag to indicate if print the information}
}
\value{
A Splice object with meta sites stored the in the meta_sites slot.
}
\description{
An interface function to apply geneSiteTable_df for target genes in a Splice object
}
\details{
An interface function to apply geneSiteTable_df for target genes in a Splice object
}
