% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{geneSiteTable}
\alias{geneSiteTable}
\title{generic geneSiteTable function definition}
\usage{
geneSiteTable(object, gene_bed, genes, eps, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene_bed}{The gene bed annotation, should include the gene name and the gene strand.}

\item{genes}{The target genes to generate meta sites, default for all genes}

\item{eps}{The threshold for two sites to be merged}

\item{...}{other possible parameters for genes_groups_GLRT.base}
}
\description{
generic geneSiteTable function definition
}
