% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{geneSiteTable_df}
\alias{geneSiteTable_df}
\title{geneSiteTable_df}
\usage{
geneSiteTable_df(
  data,
  strand,
  eps = 0.05,
  cell_col = "cell",
  iso_col = "isoform",
  count_col = "count",
  polyA_col = "polyA",
  split = "|",
  sep = ","
)
}
\arguments{
\item{data}{A dataframe recording the isoform for a gene}

\item{strand}{The strand of the target gene}

\item{eps}{The threshold for two sites to be merged}

\item{cell_col}{The name of the column recording the cell barcode}

\item{iso_col}{The name of the column recording the isoform string}

\item{count_col}{The name of the column recording the isoform count in each cell}

\item{polyA_col}{The name of the column recording the polyA existence for a read}

\item{split}{The chracter to split each exon within the isoform}

\item{sep}{The chracter to split each the start and end sites of an exon}
}
\value{
A list including two dataframes, the fisrt one records the gene count for each site in each cell,
while the second one record the spliced-in site count.
}
\description{
merge sites into meta sites and summarise the site count and gene count for each single cell
}
\details{
An integration function for spliceTable() and cellGeneSiteCount()
}
