% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{genesSitesPhi,Splice,character-method}
\alias{genesSitesPhi,Splice,character-method}
\title{generic genesSitesPhi function for Splice object}
\usage{
\S4method{genesSitesPhi}{Splice,character}(object, genes, ...)
}
\arguments{
\item{object}{the Splice object}

\item{...}{parameters for genesSitesPhi.base}
}
\description{
generic genesSitesPhi function for Splice object
}
