% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{genesSitesPhi.base}
\alias{genesSitesPhi.base}
\title{compute phi for exons in multi-genes}
\usage{
genesSitesPhi.base(
  spliceOb,
  genes,
  cells = "all",
  cell_count_thresh = 20,
  exprs_thresh = 10,
  iters = 100,
  psi_lwr = 0.1,
  psi_upr = 0.9,
  phi_conf_thresh = 0.2,
  method = match.arg("beta", "psi"),
  CI_level = 0.05,
  cores = 4,
  mode = c("sequential", "multisession", "multicore"),
  ...
)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{cells}{cells to use in building exon count table}

\item{cell_count_thresh}{the minimum number of cells expressing this gene}

\item{exprs_thresh}{the minimum gene expression to calculate the exon psi (for unbiased estimation)}

\item{iters}{iteration for bootstrap to generate the confidence interval of phi}

\item{psi_lwr, psi_upr}{The lower and upper bound to preserve as alternative spliced exons}

\item{phi_conf_thresh}{The maximum length of confidence interval to filter out inconfident
phi estimation}

\item{method}{The method to use for phi estimation}

\item{CI_level}{The alpha for the confidence interval of phi estimation}

\item{cores}{The number of cores for parallization}

\item{mode}{The mode of parallization}

\item{...}{Other parameters for geneSitesPsi.base}
}
\value{
a dataframe recording the phi estimation
}
\description{
compute phi for exons in a gene from a Splice object
}
