% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{genesSitesPhi}
\alias{genesSitesPhi}
\title{generic genesSitesPhi function definition}
\usage{
genesSitesPhi(object, genes, ...)
}
\arguments{
\item{object}{the Splice object}

\item{genes}{vector of names of the target gene}

\item{...}{other possible parameters}
}
\description{
generic genesSitesPhi function definition
}
