% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getIsoform,Seurat,character-method}
\alias{getIsoform,Seurat,character-method}
\title{generic getIsoform function for Seurat object}
\usage{
\S4method{getIsoform}{Seurat,character}(object, genes, slot = "splice", ...)
}
\arguments{
\item{object}{a Seurat object}

\item{genes}{A vector of genes to extract isoforms}

\item{slot}{the slot name to store Splice object in Seurat object}

\item{...}{parameters for getIsoform.base}
}
\value{
A dataframe which store the expression and related information for each
isoform.
}
\description{
extract isoforms for multigenes from selected cells for Seurat object
}
