% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getIsoform,Splice,character-method}
\alias{getIsoform,Splice,character-method}
\title{generic getIsoform function for Splice object}
\usage{
\S4method{getIsoform}{Splice,character}(object, genes, ...)
}
\arguments{
\item{object}{a Splice object}

\item{genes}{A vector of genes to extract isoforms}

\item{...}{parameters for getIsoform.base}
}
\value{
A dataframe which store the expression and related information for each
isoform.
}
\description{
extract isoforms for multigenes from selected cells for Splice object
}
