% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getIsoform.base}
\alias{getIsoform.base}
\title{extract isoforms}
\usage{
getIsoform.base(spliceob, genes, cells = "all")
}
\arguments{
\item{spliceob}{A Splice object}

\item{genes}{A vector of genes to extract isoforms}

\item{cells}{A vector of cells to choose to extract isoforms. Default to choose
all cells}
}
\value{
A dataframe which store the expression and related information for each
isoform.
}
\description{
extract isoforms for multigenes from selected cells
}
