% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getIsoform}
\alias{getIsoform}
\title{generic getIsoform function definition}
\usage{
getIsoform(object, genes, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{genes}{A vector of genes to extract isoforms}

\item{...}{other possible parameters for getIsoform.base}
}
\description{
generic getIsoform function definition
}
