% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getMetaSites}
\alias{getMetaSites}
\title{getMetaSites}
\usage{
getMetaSites(spliceOb, gene, cells = "all", verbose = TRUE)
}
\arguments{
\item{spliceOb}{the Splice object}

\item{gene}{name of the target gene}

\item{cells}{cells to use in building exon count table}

\item{verbose}{The flag to indicate if output message}
}
\value{
a metaSite object
}
\description{
Extract the metaSites object for a gene for a subset of cells from the splice object
}
