% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{getSplice}
\alias{getSplice}
\title{extract the Splice object from the Seurat object}
\usage{
getSplice(object, slot = "splice")
}
\arguments{
\item{object}{A seurat object with a Splice object embedded in}

\item{slot}{the slot name to store Splice object in Seurat object}
}
\value{
A Splice object.
}
\description{
extract the Splice object from the Seurat object
}
