% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{GLRT}
\alias{GLRT}
\title{Generalized likelihood ratio test.}
\usage{
GLRT(
  exon_count1,
  gene_count1,
  exon_count2,
  gene_count2,
  iters = 100,
  psi_num = 500,
  p_thresh = 0.05
)
}
\arguments{
\item{exon_count1}{a vector of exon count for cell population 1}

\item{gene_count1}{a vector of gene count for cell population 1}

\item{exon_count2}{a vector of exon count for cell population 2}

\item{gene_count2}{a vector of gene count for cell population 2}

\item{iters}{the number of iterations for bootstrap}

\item{psi_num}{the sample size from two beta distributions to calculate the earth-moving distance}

\item{p_thresh}{the threshold for the p-value}
}
\value{
a vector of p-value and parameters for the psi distributions in two
cell populations if they have significant difference
}
\description{
Using beta-binomial to model the exon count and do
generalized likelihood ratio test for two cell populations to test if they have the
same psi distribution.
}
