% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{gtf_bed_corres}
\alias{gtf_bed_corres}
\title{assign gtf exon annotations to self annotated sub-exons}
\usage{
gtf_bed_corres(
  gtf,
  bed,
  gene,
  gtf_gene_col = "gene_id",
  bed_gene_col = "gene",
  transcript_col = "transname",
  exon_id = NULL,
  strand_col = "strand",
  gtf_start_col = "start",
  gtf_end_col = "end",
  bed_start_col = "start",
  bed_end_col = "end"
)
}
\arguments{
\item{gtf}{a dataframe of gtf annotation}

\item{bed}{a dataframe of self annotated gene bed}

\item{gene}{the gene to be be matched}

\item{gtf_gene_col}{the colname in the gtf which indicates the gene id}

\item{bed_gene_col}{the colname in the bed which indicates the gene id}

\item{transcript_col}{the colname in the gtf which indicates the name of transcripts}

\item{exon_id}{A vector of string to name the self annotated exon, should have the same length
as the self-annotated exons}

\item{strand_col}{The colname to indicate the strand in the self-annotated gene bed}

\item{gtf_start_col, gtf_end_col}{The colnames for start and end sites of the canonical exons
in the gtf annotation}

\item{bed_start_col, bed_end_col}{The colnames for start and end sites of the canonical exons
in the bed annotation}
}
\value{
A list of sub-exon string named by isoforms in the gtf.
}
\description{
assign gtf exon annotations for a gene to self annotated sub-exons
}
