% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{HighExprsGene,Seurat,numeric-method}
\alias{HighExprsGene,Seurat,numeric-method}
\title{generic HighExprsGene function for Seurat object}
\usage{
\S4method{HighExprsGene}{Seurat,numeric}(object, thresh = 100, slot = "RNA")
}
\arguments{
\item{object}{the Seurat object}

\item{thresh}{the minimum total gene expression}

\item{slot}{the slot name RNA expression assay in the Seurat object}
}
\description{
choose genes with high total expression for phi calculation
}
