% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{HighExprsGene}
\alias{HighExprsGene}
\title{generic HighExprsGene function definition}
\usage{
HighExprsGene(object, thresh, ...)
}
\arguments{
\item{object}{the Seurat object}

\item{thresh}{the threshold for minimum total gene expression}

\item{...}{other possible parameters}
}
\description{
generic HighExprsGene function definition
}
