% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{ifMerge}
\alias{ifMerge}
\title{judge if two exons can be merged}
\usage{
ifMerge(a, b)
}
\arguments{
\item{a}{A numerical vector for the existence of an exon in each isoform}

\item{b}{A numerical vector for the existence of an exon in each isoform}
}
\value{
TRUE or FALSE, indicating if the two exons should be merged
}
\description{
judge if two exons can be merged
}
\details{
If two exons are co-existing or mutually exclusive, they have the same information,
and can considered as one exon in downstream analysis
}
