% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{isoform2bases}
\alias{isoform2bases}
\title{isoform2bases}
\usage{
isoform2bases(isoform, count = 1, sep = ",", split = "|")
}
\arguments{
\item{isoform}{The isoform sequence, should be formatted as "s1,e1|s2,e2|...", s means the start site of an exon
while e means the end site.}

\item{count}{The count for each base}

\item{sep}{The character to split the start and end sites of an exon.}

\item{split}{The character to split the exon in the isoform sequence.}
}
\value{
a vector of bases
}
\description{
transform an isoform sequence into a vector of base this isoform covered
}
