% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{isoform2sites}
\alias{isoform2sites}
\title{isoform2sites}
\usage{
isoform2sites(isoform, split = "|", sep = ",")
}
\arguments{
\item{isoform}{A string representing the isoform}

\item{split}{The chracter to split each exon within the isoform}

\item{sep}{The chracter to split each the start and end sites of an exon}
}
\value{
A string vector, each element is a splicing site
}
\description{
split the isoform string to a vector of splicing sites
}
\details{
split the isoform string to a vector of splicing sites
}
