% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{junction_count}
\alias{junction_count}
\title{junction_count}
\usage{
junction_count(isoform, count = 1, polyA = 0, sep = ",", split = "|")
}
\arguments{
\item{isoform}{The isoform sequence, should be formatted as "s1,e1|s2,e2|...", s means the start site of an exon
while e means the end site.}

\item{count}{The count for each base}

\item{polyA}{A flag to indicate if the read has a polyA tail}

\item{sep}{The character to split the start and end sites of an exon.}

\item{split}{The character to split the exon in the isoform sequence.}
}
\value{
A matrix recording the junction count, the first column is the junction
identity and the second column is the count
}
\description{
collect the splice junction in an isoform sequence
}
