% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom_mle.R
\name{ll}
\alias{ll}
\title{ll}
\usage{
ll(exon_count, n, alpha, beta)
}
\arguments{
\item{exon_count}{a vector of exon count}

\item{n}{a vector of gene count with the same length as exon_count}

\item{alpha}{alpha for beta distribution}

\item{beta}{beta for beta distribution}
}
\value{
a numerical value as negative log likelihood
}
\description{
compute log likelihood for exon count and gene count given a
beta-binomial distribution
}
