% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\docType{class}
\name{metaSite-class}
\alias{metaSite-class}
\title{The metaSite Class}
\description{
The metaSite class is an object to store the meta splicing sites information. It includes three slots, the gene
count and spliced-in count per cell for each meta site the orginal splicing sites that each meta site corresponds to.
}
\section{Slots}{

\describe{
\item{\code{cellGeneCount}}{A cell by meta site matrix, recording the gene expression count.}

\item{\code{cellSiteCount}}{A cell by meta site matrix, recording the spliced-in count.}

\item{\code{sites}}{A list with length as the number of meta splicing sites, Each element records the corresponding original splicing sites.}
}}

\concept{objects}
