% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\name{metasite_retrieve}
\alias{metasite_retrieve}
\title{metasite_retrieve}
\usage{
metasite_retrieve(splice_ob, genes, sites)
}
\arguments{
\item{genes}{A vector of genes to extract isoforms}

\item{sites}{A vector of meta splicing sites, usually has format as "metasite_id"}

\item{spliceob}{A Splice object}
}
\value{
A vector of string, each element is the pasted splicing sites for a meta site.
}
\description{
recover the metasites to the corresponding original sites for multiple genes
}
