% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{metaSpliceSite}
\alias{metaSpliceSite}
\title{metaSpliceSite}
\usage{
metaSpliceSite(sites_table, count, eps = 0.05)
}
\arguments{
\item{sites_table}{The site count table output from sites2matrix()}

\item{eps}{The threshold for two sites to be merged}
}
\value{
A list with two elements, the first one is the count matrix of meta splicing sites,
the second one is the component of the meta splicing sites
}
\description{
Merge sites into meta sites based on their co-existence or if they are mutually exclusive
}
\details{
Merge sites into meta sites based on their co-existence or if they are mutually exclusive
}
