% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom_mle.R
\name{mle_bb}
\alias{mle_bb}
\title{mle_bb}
\usage{
mle_bb(exon_count, gene_count, start = NULL)
}
\arguments{
\item{exon_count}{a vector of exon count}

\item{gene_count}{a vector of gene count with the same length as exon_count}

\item{start}{a list of start points, each line should be a pair of alpha and beta}
}
\value{
a list of result for mle
}
\description{
mle for beta-binomial with multiple start points to avoid local optimization
}
