% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binom_mle.R
\name{mle_module}
\alias{mle_module}
\title{mle_module}
\usage{
mle_module(exon_count, gene_count, start = NULL)
}
\arguments{
\item{exon_count}{a vector of exon count}

\item{gene_count}{a vector of gene count with the same length as exon_count}

\item{start}{a list of start points, each line should be a pair of alpha and beta}
}
\value{
a mle2 object
}
\description{
MLE estimation for a beta-binomial distribution with different number of total experiments
}
