% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{moduleExtract}
\alias{moduleExtract}
\title{moduleExtract}
\usage{
moduleExtract(coef_matrix, eps = 0.05)
}
\arguments{
\item{coef_matrix}{A square matrix of logical values, indicating if each two sites should be merged}

\item{eps}{The threshold for two sites to be merged}
}
\value{
A list of vectors with variable length,each vector represents a group of sites to be merged
}
\description{
extract exons to be merged according to the exon merge flag table
}
\details{
sites which can be merged according to the exon merge flag table are extracted
together into a group. All different merge groups are collected as a list
}
