% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{multiExonsMerge}
\alias{multiExonsMerge}
\title{merge multi-exons}
\usage{
multiExonsMerge(data)
}
\arguments{
\item{data}{a dataframe of count for mutil-exons in each isoform}
}
\value{
a count vector for merged exon
}
\description{
merge the existence of multi-exons in each isoform
}
\details{
Input a dataframe of count for mutil-exons in each isoform, mering them by pair iteratively
}
