% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{phi}
\alias{phi}
\title{count exon phi}
\usage{
phi(x)
}
\arguments{
\item{x}{a vector of exon psi across cells}
}
\value{
a numeric value for exon phi
}
\description{
calculate the alternative splicing heterogeneity for an exon
}
