% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{phi_conf}
\alias{phi_conf}
\title{count the confidence interval for exon phi}
\usage{
phi_conf(x, iters = 100)
}
\arguments{
\item{x}{a vector of exon psi across cells}

\item{iters}{the number of sampling iteration in bootstrap}
}
\value{
a numeric value for exon phi
}
\description{
using bootstrap to calculate the confidence interval for exon phi
}
