% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi_conf_bootstrap}
\alias{phi_conf_bootstrap}
\title{phi_conf_bootstrap}
\usage{
phi_conf_bootstrap(x, iters = 100)
}
\arguments{
\item{x}{a vector of exon psi across cells}

\item{iters}{the number of sampling iteration in bootstrap}
}
\value{
a numeric vector of the confidence interval
}
\description{
using bootstrap to calculate the confidence interval for exon phi
}
