% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi_conf_delta}
\alias{phi_conf_delta}
\title{phi_conf_delta}
\usage{
phi_conf_delta(alpha, beta, covmat, a = 0.05)
}
\arguments{
\item{alpha}{alpha parameter for beta distribution}

\item{beta}{beta parameter for beta distribution}

\item{covmat}{The covariance matrix for alpha and beta}

\item{a}{The range of the confidence interval}
}
\value{
a numeric vector of the confidence interval
}
\description{
using delta method to estimate the confidence interval for exon phi
}
