% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi_from_beta}
\alias{phi_from_beta}
\title{phi_from_beta}
\usage{
phi_from_beta(alpha, beta)
}
\arguments{
\item{alpha}{alpha parameter for beta distribution}

\item{beta}{beta parameter for beta distribution}
}
\value{
a numeric value for exon phi
}
\description{
phi estimation from beta distribution
}
