% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi_from_psi}
\alias{phi_from_psi}
\title{phi_from_psi}
\usage{
phi_from_psi(x)
}
\arguments{
\item{x}{a vector of exon psi across cells}
}
\value{
a numeric value for exon phi
}
\description{
calculate the alternative splicing heterogeneity for an exon based on psi estimation
}
