% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phiNabla}
\alias{phiNabla}
\title{phiNabla}
\usage{
phiNabla(alpha, beta)
}
\arguments{
\item{alpha}{alpha parameter for beta distribution}

\item{beta}{beta parameter for beta distribution}
}
\value{
a numerical vector for the first derivative for alpha and beta
}
\description{
The Nabla for phi estimation from alpha and beta
}
