% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiCellPlot,Seurat,character,character,character-method}
\alias{psiCellPlot,Seurat,character,character,character-method}
\title{generic psiCellPlot function for Seurat object}
\usage{
\S4method{psiCellPlot}{Seurat,character,character,character}(object, gene, exons, cell_embedding = "umap", slot = "splice", ...)
}
\arguments{
\item{object}{the Seurat object with both cell embeddings and Splice object}

\item{gene}{the name of specified gene}

\item{exons}{the selected exon to be plot}

\item{cell_embedding}{the name of embedding to use from the Seurat object}

\item{slot}{the slot name to store Splice object in Seurat object}

\item{...}{parameters for gene_exons_psi.base to calculate psi value}
}
\value{
a ggplot object of psi in cell embedding
}
\description{
plot psi value in cell embeddings for a Seurat object
}
