% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiCellPlot.base}
\alias{psiCellPlot.base}
\title{plot psi value in cell embeddings}
\usage{
psiCellPlot.base(spliceOb, gene, exons, cell_embedding, dims = c(1, 2), ...)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{exons}{the selected exon to be plot}

\item{cell_embedding}{A dataframe of cell embedding coordinates}

\item{dims}{Two dimensions to use to plot cell embedding}

\item{...}{parameters for gene_exons_psi.base to calculate psi value}
}
\value{
a ggplot object of psi in cell embedding
}
\description{
plot psi value in cell embeddings
}
