% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiCellPlot}
\alias{psiCellPlot}
\title{generic psiCellPlot function definition}
\usage{
psiCellPlot(object, gene, exons, cell_embedding, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene}{the name of specified gene}

\item{exons}{the selected exon to be plot}

\item{cell_embedding}{A dataframe of cell embedding coordinates}

\item{...}{parameters for psiCellPlot.base}
}
\description{
generic psiCellPlot function definition
}
