% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{psiCellPlotData}
\alias{psiCellPlotData}
\title{psiCellPlotData}
\usage{
psiCellPlotData(spliceOb, gene, sites, cells = "all", exprs_thresh = 10, ...)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{sites}{the selected exon to be plot}

\item{cells}{the cells to be included in the psi plot}

\item{exprs_thresh}{the minimum gene expression required for psi calculation}

\item{...}{parameters for geneSitesPsi.base to calculate psi value}
}
\value{
a list including the gene count and the psi for the target splicing site to be plotted
}
\description{
Prepare the plot data for psiCellPlot
}
