% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiHist,Seurat,character,character-method}
\alias{psiHist,Seurat,character,character-method}
\title{generic psiHist function for Seurat object}
\usage{
\S4method{psiHist}{Seurat,character,character}(
  object,
  gene,
  exon,
  cells = "all",
  slot = "splice",
  group_col = NULL,
  group = NULL,
  ...
)
}
\arguments{
\item{object}{The Seurat object}

\item{gene}{the name of specified gene}

\item{exon}{the selected exon to be plot}

\item{cells}{the cells to be included in the psi plot}

\item{slot}{the slot name to store Splice object in Seurat object}

\item{group_col}{if group_col and group is not NULL, parameters cells will be ignored
and cells within the group will be selected for psi plot}

\item{group}{cell group in meta.data of Seurat}

\item{...}{aesthetic parameters for ggplot}
}
\value{
a ggplot object of psi distribution histogram
}
\description{
plot psi distribution for a cell group into histogram for a Seurat object
}
