% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiHist.base}
\alias{psiHist.base}
\title{plot psi distribution}
\usage{
psiHist.base(
  spliceOb,
  gene,
  exon,
  cells = "all",
  gene_bed = NULL,
  gtf = NULL,
  alpha = NULL,
  beta = NULL,
  exprs_thresh = 5,
  exon_len_thresh = 10,
  sep = "|",
  binwidth = 0.05,
  low = "steelblue",
  high = "coral",
  mid = "whitesmoke",
  midpoint = NULL,
  ...
)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{exon}{the selected exon to be plot}

\item{cells}{the cells to be included in the psi plot}

\item{gene_bed}{a dataframe of gene_bed annotation}

\item{gtf}{a dataframe of gtf annotation}

\item{alpha}{the parameter alpha for fitted beta distribution for psi}

\item{beta}{the parameter beta for fitted beta distribution for psi}

\item{exprs_thresh}{the minimum gene expression required for psi calculation}

\item{exon_len_thresh}{the minimum length of exons for GLRT}

\item{sep}{the character to split exons in the isoform string}

\item{binwidth}{the bin size for the histogram ranging from (0,1]}

\item{low, high, mid}{Colors for the bins to represent the mean gene count in each bin}

\item{midpoint}{the gene count thresh to set the middle color}

\item{...}{aesthetic parameters for ggplot}
}
\value{
a ggplot object of psi distribution histogram
}
\description{
plot psi distribution for a cell group into histogram
}
