% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{range_scale}
\alias{range_scale}
\title{range_scale}
\usage{
range_scale(data, lwr = 1, upr = 5)
}
\arguments{
\item{data}{a numeric vector}

\item{lwr, upr}{the lower/upper bound of the range}
}
\value{
a numeric vector recording the scaled data
}
\description{
scale the data into a fixed range
}
