% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{sashimi}
\alias{sashimi}
\title{sashimi}
\usage{
sashimi(spliceOb, gene, cells, ...)
}
\arguments{
\item{spliceOb}{The input splice object}

\item{gene}{The target gene}

\item{cells}{The target cell population}

\item{...}{Omitted parameters for aesthetic  parameters in sashimi_plot}
}
\value{
a ggplot object.
}
\description{
generate the sashimi plot for a gene from the Splice object
}
