% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{sashimi_plot}
\alias{sashimi_plot}
\title{sashimi_plot}
\usage{
sashimi_plot(
  coverage,
  junction,
  filter_ratio = 20,
  color_id = 1,
  region = NULL,
  lwr = 1,
  upr = 5,
  color_set = NULL
)
}
\arguments{
\item{coverage}{The base count}

\item{junction}{The junction count}

\item{filter_ratio}{Threshold to filter out lowly covered base}

\item{color_id}{The prdefined color set to use, currently there are two sets}

\item{region}{The xlab annotation for the sashimi plot}

\item{lwr, upr}{the lower/upper bound of the range}

\item{color_set}{a vector to record user defined color set to use, should include two colors, one for the
base coverage and the other for the junction connection.}
}
\value{
a ggplot object.
}
\description{
generate the sashimi plot for the isoform
}
