% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{sashimi_plot_data}
\alias{sashimi_plot_data}
\title{sashimi_plot_data}
\usage{
sashimi_plot_data(isoforms, counts, polyA = NULL, sep = ",", split = "|")
}
\arguments{
\item{polyA}{A flag to indicate if the read has a polyA tail}

\item{sep}{The character to split the start and end sites of an exon.}

\item{split}{The character to split the exon in the isoform sequence.}
}
\value{
A list recording the base count and the junction count.
}
\description{
transform the isoform count data to the format for sahimi plot
}
