% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{siteMerge}
\alias{siteMerge}
\title{siteMerge}
\usage{
siteMerge(site_matrix)
}
\arguments{
\item{site_matrix}{A binary matrix to indicate if the sites exist}
}
\value{
A vector for the meta splicing site.
}
\description{
Merge sites within the same group
}
\details{
Merge sites within the same group
}
