% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{sites2matrix}
\alias{sites2matrix}
\title{sites2matrix}
\usage{
sites2matrix(data, in_sites_index, out_sites_index, strand)
}
\arguments{
\item{data}{A dataframe recording the splicing sites for a read, including start, middle splicing sites, end
and polyA}

\item{in_sites_index}{A string vector recording unique start position for all middle exons}

\item{out_sites_index}{A string vector recording unique end position for all middle exons}

\item{strand}{The strand of this gene}
}
\value{
A binary matrix indicating the splicing sites existence for all reads from a gene
}
\description{
convert the splicing sites vector for all reads to a binary matrix
}
\details{
This function applies sites2vec to all reads and concat all vector results to a matrix
}
