% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_table.R
\name{sites_coef_matrix}
\alias{sites_coef_matrix}
\title{sites_coef_matrix}
\usage{
sites_coef_matrix(sites_table, count)
}
\arguments{
\item{sites_table}{The site count table output from sites2matrix()}
}
\value{
A square matrix of logical values, indicating if each two exons should be merged
}
\description{
judge if each site in the gene can be merged with other
}
\details{
Iteratively check if each pair of exons in the gene co-existing or
mutually exclusive
}
