% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceOb.R
\docType{class}
\name{Splice-class}
\alias{Splice-class}
\title{The Splice Class}
\description{
The Splice class is an intermediate data storage class that stores the isoforms and other
related information needed for performing downstream analyses: including highly variable
exons detection and differential alternative splicing analysis
}
\section{Slots}{

\describe{
\item{\code{cells}}{The vector to store cell names}

\item{\code{genes}}{The vector to store gene names.}

\item{\code{isoforms}}{The vector to store isoform names.}

\item{\code{count}}{A data frame to store the expression, polyA and other related information
for each isoform in each cell}
}}

\concept{objects}
