% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{transcript_corres}
\alias{transcript_corres}
\title{assign canonical isoform to self annotated sub-exons}
\usage{
transcript_corres(
  starts,
  ends,
  bed,
  exon_id = NULL,
  strand_col = "strand",
  start_col = "start",
  end_col = "end"
)
}
\arguments{
\item{starts, ends}{The start and end sites of the canonical exons within the isoform}

\item{bed}{the gene bed annotation}

\item{exon_id}{A vector of string to name the self annotated exon, should have the same length
as the self-annotated exons}

\item{strand_col}{The colname to indicate the strand in the self-annotated gene bed}

\item{start_col, end_col}{The colnames of start and end sites in the self-annotated gene bed}
}
\value{
A string of sub-exons within the isoform, seperated by "|".
}
\description{
assign canonical isoform (represented by multi canonical exons) to
self annotated sub-exons
}
