% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRT_functions.R
\name{GLRT_sig_plot}
\alias{GLRT_sig_plot}
\title{plot for GLRT significant results}
\usage{
GLRT_sig_plot(
  data,
  gene_col = "gene",
  mean_diff = "mean_diff",
  var_diff = "var_diff",
  q = "q",
  q_thresh = 3,
  q_mid = NULL,
  pt.size = 2,
  text.size = 3,
  low = "steelblue",
  mid = "whitesmoke",
  high = "tomato"
)
}
\arguments{
\item{data}{the GLRT result, a dataframe output from genes_groups_GLRT()}

\item{gene_col}{The name of column in the data which stores the gene name}

\item{mean_diff}{The name of column in the data which stores the mean difference
of the psi distributions}

\item{var_diff}{The name of column in the data which stores the variance difference
of the psi distributions}

\item{q}{The name of column in the data which stores the q value}

\item{q_thresh}{the minimum of q to show the gene name in the scatter plot}

\item{q_mid}{The scatter plot is colored by q value, q_mid is set for the middle color point}

\item{pt.size}{The point size in the scatter}

\item{text.size}{The size of the gene name annotation on the point}

\item{low, mid, high}{The color to indicate the q value}
}
\value{
a ggplot object
}
\description{
plot the scatter plot for mean change versus variance change of exon psi
between two cell groups
}
