% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_table.R
\name{moduleExtract}
\alias{moduleExtract}
\title{collect exons to be merged into a list}
\usage{
moduleExtract(data)
}
\arguments{
\item{data}{A square matrix of logical values, indicating if each two exons should be merged}
}
\value{
A list of vectors with variable length,each vector represents a group of exons to be merged
}
\description{
extract exons to be merged according to the exon merge flag table
}
\details{
exons which can be merged according to the exon merge flag table are extracted
together into a group. All different merge groups are collected as a list
}
