% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{psiCellPlot.base}
\alias{psiCellPlot.base}
\title{psiCellPlot.base}
\usage{
psiCellPlot.base(
  spliceOb,
  gene,
  sites,
  cell_embedding,
  dims = c(1, 2),
  cells = "all",
  exprs_thresh = 10,
  ...
)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{sites}{the selected exon to be plot}

\item{cell_embedding}{A dataframe of cell embedding coordinates}

\item{dims}{Two dimensions to use to plot cell embedding}

\item{cells}{the cells to be included in the psi plot}

\item{exprs_thresh}{the minimum gene expression required for psi calculation}

\item{...}{parameters for gene_exons_psi.base to calculate psi value}
}
\value{
a ggplot object of psi in cell embedding
}
\description{
plot psi value in cell embeddings
}
