% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiHist,Splice,character,character-method}
\alias{psiHist,Splice,character,character-method}
\title{generic psiHist function for Splice object}
\usage{
\S4method{psiHist}{Splice,character,character}(object, gene, exon, ...)
}
\arguments{
\item{object}{The Splice object}

\item{gene}{the name of specified gene}

\item{exon}{the selected exon to be plot}

\item{...}{aesthetic parameters for ggplot}
}
\value{
a ggplot object of psi distribution histogram
}
\description{
plot psi distribution for a cell group into histogram for a Splice object
}
