% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_visual.R
\name{psiHist.base}
\alias{psiHist.base}
\title{plot psi distribution}
\usage{
psiHist.base(
  spliceOb,
  gene,
  site,
  cells = "all",
  alpha = NULL,
  beta = NULL,
  exprs_thresh = 5,
  binwidth = 0.05,
  low = "steelblue",
  high = "coral",
  mid = "whitesmoke",
  midpoint = NULL,
  binsize = 8
)
}
\arguments{
\item{spliceOb}{a Splice object}

\item{gene}{the name of specified gene}

\item{site}{the selected site to be plot}

\item{cells}{the cells to be included in the psi plot}

\item{alpha}{the parameter alpha for fitted beta distribution for psi}

\item{beta}{the parameter beta for fitted beta distribution for psi}

\item{exprs_thresh}{the minimum gene expression required for psi calculation}

\item{binwidth}{the bin size for the histogram ranging from (0,1]}

\item{low, high, mid}{Colors for the bins to represent the mean gene count in each bin}

\item{midpoint}{the gene count thresh to set the middle color}

\item{...}{aesthetic parameters for ggplot}
}
\value{
a ggplot object of psi distribution histogram
}
\description{
plot psi distribution for a cell group into histogram
}
