% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_phi.R
\name{psiHist}
\alias{psiHist}
\title{generic psiHist function definition}
\usage{
psiHist(object, gene, exon, ...)
}
\arguments{
\item{object}{the Splice or Seurat object}

\item{gene}{the name of specified gene}

\item{exon}{the selected exon to be plot}

\item{...}{parameters for psiHist.base}
}
\description{
generic psiHist function definition
}
